Conflicts:
	src/main/java/kr/or/member/dao/Dmdao.java
	src/main/resources/mapper/dmSQL.xml
	src/main/webapp/WEB-INF/views/member/dmMain.jsp

+++ src/main/java/kr/or/fund/model/vo/TmpFundCalculate.java
@@ -17,7 +17,7 @@ public class TmpFundCalculate {
 	private String tfcRepBirth;
 	private String tfcMail;
 	private String tfcBankName;
-	private int tfcBankNo;
+	private long tfcBankNo;
 	private String tfcBankAccount;
 	private String tfcBankFilepath;
